requirements.txt
absl-py==0.13.0
appdirs==1.4.4
arabic-reshaper==2.1.3
asgiref==3.3.4
astroid==2.5.6
astunparse==1.6.3
backcall==0.2.0
CacheControl==0.12.6
cachetools==4.2.2
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
colorama==0.4.4
cryptography==3.4.7
cycler==0.10.0
decorator==5.0.9
defusedxml==0.7.1
Deprecated==1.2.12
distlib==0.3.1
dj-database-url==0.5.0
Django==3.2
django-cleanup==5.2.0
django-crispy-forms==1.11.2
django-filter==2.4.0
django-heroku==0.3.1
django-js-asset==1.2.2
django-mptt==0.12.0
django-widget-tweaks==1.4.8
docutils==0.17.1
filelock==3.0.12
firebase-admin==5.0.1
flatbuffers==1.12
future==0.18.2
gast==0.4.0
gcloud==0.18.3
google-api-core==1.31.0
google-api-python-client==2.14.0
google-auth==1.32.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.4
google-cloud-core==1.7.1
google-cloud-firestore==2.2.0
google-cloud-storage==1.41.1
google-crc32c==1.1.2
google-pasta==0.2.0
google-resumable-media==1.3.1
googleapis-common-protos==1.53.0
grpcio==1.34.1
gunicorn==20.1.0
h5py==3.1.0
html5lib==1.1
httplib2==0.19.1
idna==2.10
ipykernel==5.5.5
ipython==7.24.1
ipython-genutils==0.2.0
isort==5.8.0
jedi==0.18.0
joblib==1.0.1
jupyter-client==6.1.12
jupyter-core==4.7.1
jwcrypto==0.9.1
jws==0.1.3
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
kivy-deps.angle==0.3.0
kivy-deps.glew==0.3.0
kivy-deps.sdl2==0.3.1
Kivy-Garden==0.1.4
kivymd==0.104.2
kiwisolver==1.3.1
lazy-object-proxy==1.6.0
lxml==4.6.3
Markdown==3.3.4
matplotlib==3.4.2
matplotlib-inline==0.1.2
mccabe==0.6.1
msgpack==1.0.2
mysql-connector-python==8.0.25
mysqlclient==2.0.3
numpy==1.19.5
oauth2client==4.1.3
oauthlib==3.1.0
opt-einsum==3.3.0
packaging==21.0
pandas==1.2.4
pandas-datareader==0.9.0
parso==0.8.2
pickleshare==0.7.5
Pillow==8.2.0
prompt-toolkit==3.0.19
proto-plus==1.19.0
protobuf==3.17.0
psycopg2==2.9.1
pushbullet.py==0.9.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pycparser==2.20
pycryptodome==3.10.1
Pygments==2.9.0
PyJWT==2.1.0
pylint==2.8.2
pylint-django==2.4.4
pylint-plugin-utils==0.6
pymongo==3.11.4
pyparsing==2.4.7
PyPDF2==1.26.0
pypiwin32==223
Pyrebase4==4.5.0
pyserial==3.5
python-bidi==0.4.2
python-dateutil==2.8.1
python-decouple==3.4
python-http-client==3.3.2
python-jwt==2.0.1
python-magic==0.4.24
python3-openid==3.2.0
pytz==2021.1
pywin32==301
pyzmq==22.1.0
reportlab==3.5.67
requests==2.25.1
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
rsa==4.7.2
scikit-learn==0.24.2
scipy==1.6.3
seaborn==0.11.1
six==1.15.0
sklearn-pandas==2.2.0
sqlparse==0.4.1
sseclient==0.0.27
starkbank-ecdsa==1.1.0
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.5.0
tensorflow-estimator==2.5.0
termcolor==1.1.0
threadpoolctl==2.1.0
toml==0.10.2
tornado==6.1
traitlets==5.0.5
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.26.4
virtualenv==20.4.4
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.1.0
Werkzeug==2.0.1
whitenoise==5.2.0
wrapt==1.12.1
